RYP-59 Added Dockerfile for running the service.
After looking at potentially including the build process in the Docker image, it was clear that to do so (the plan is to share logic and DTO classes between the services where sensible) would require at least four Docker files: one for the build at the root, then one for each service. Combined with the plans to add sonar coverage and several other items that are not part of the default build that would be included in Docker. For now, we are leaving the build logic to the Gitlab CI file and considering a dedicated Docker build image later.