RYP-107 Started to provide API to view announcements by ID.
* RYP-107 Added basic support to view announcements on the website by announcement ID

Added enhanced information to the AS 2.0 blob stored alongside the announcement metadata, like the project on which behalf the announcement is made, as well as URL and title etc. Most of it is server side rendering.

* RYP-107 Started to provide API to view announcements by ID.

Still need to do some work on the demo site to show the content and link to the project.