RYP-146 Implemented core logic around CIP-0022 verification with support for blockfrost and db-sync
* RYP-146 Finalizing remaining issues in pool verification frontend and backend, as well as verification persistence

Updated OpenAPI and clients to match latest configurations. Also renamed another class to clarify its DTO purpose. Also adjusted docker files and docker compose file samples.

* RYP-146 Adding MongoDB capabilities to verification to store pool verifications.

* RYP-146 Renaming StakepoolVerification class to clarify its Dto purpose

* RYP-146 Fixed broken test

Still missing the storage of pool verifications.

* RYP-146 Added frontend logic to create an SPO project

Does all the verification through the frontend and confirms on the backend. Marks new project as verified and implement complex frontend flow for new SPOs that includes CIP-0022 steps.

* RYP-146 Implemented core logic around CIP-0022 verification with support for blockfrost and db-sync

Customized a docker image to build with LibSodium and leveraging the library and the CIP-0022 documentation to create a nonce-based challenge that expires after 5 minutes and is stored in Redis for the duration. Added custom exceptions for various failure possibilities, CBOR parsing to the core verification service and ensuring that only VRF Vkeys are expected, so that SKeys are not accidentally persisted/sent around more than needed.